requirements.txt
aiogram==3.10.0
aiohttp==3.9.5
tinkoff-investments==0.2.0b117
python-dotenv==1.0.0
pandas==2.2.0
python-dateutil==2.8.2
python-telegram-bot>=20.7
requests>=2.31.0
pip install certifi